requirements.txt
cpprb==11.0.0
click==8.1.7
einops==0.8.0
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
hydra-ax-sweeper==1.1.5
imageio==2.9.0
munch==4.0.0
pandas==2.2.2
torch==2.4.0
pyyaml==6.0.2
